Implement Menu Component
